+++ ovxlib/src/kernel/evis/pre_process_rgb_evis.c
@@ -176,11 +176,11 @@ DEF_KERNEL_INITIALIZER(_pre_process_rgb_initializer)
     {
         if (attr[0]->dfp.fl > 0)
         {
-            outputScale = (float)(1 << attr[0]->dfp.fl);
+            outputScale = (float)((int64_t)1 << attr[0]->dfp.fl);
         }
         else
         {
-            outputScale = (1.0f / (float)(1 << -attr[0]->dfp.fl));
+            outputScale = (1.0f / (float)((int64_t)1 << -attr[0]->dfp.fl));
         }
         outputZP = 0;
     }
